Add files via upload
Business Context 

We are all aware that accidents are prone everywhere due to negligent driving or climatic conditions. An insurance company always needs to be prepared to estimate the number of accidents and the claims that they can receive at a given point time. Also understanding the pattern of claims would help the companies to frame different types of policies for the users providing better benefits and at the same time increasing the premium to the company.